reordered the folder structure to minimize code repetitions
