+++ client/src/pages/Accounts/AuthForm/SignupForm/SignUpForm.tsx
@@ -1,20 +1,22 @@
-import Typography from '@material-ui/core/Typography';
-import { CircularProgress } from '@material-ui/core';
 import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
 import Box from '@material-ui/core/Box';
+import Typography from '@material-ui/core/Typography';
+import { CircularProgress } from '@material-ui/core';
 import { Formik, FormikHelpers } from 'formik';
 import * as Yup from 'yup';
-import useStyles from './useStyles';
+import useStyles from '../useStyles';
 
 interface Props {
   handleSubmit: (
     {
+      username,
       email,
       password,
     }: {
       email: string;
       password: string;
+      username: string;
     },
     {
       setStatus,
@@ -22,11 +24,12 @@ interface Props {
     }: FormikHelpers<{
       email: string;
       password: string;
+      username: string;
     }>,
   ) => void;
 }
 
-export default function Login({ handleSubmit }: Props): JSX.Element {
+const SignUpForm = ({ handleSubmit }: Props): JSX.Element => {
   const classes = useStyles();
 
   return (
@@ -34,8 +37,10 @@ export default function Login({ handleSubmit }: Props): JSX.Element {
       initialValues={{
         email: '',
         password: '',
+        username: '',
       }}
       validationSchema={Yup.object().shape({
+        username: Yup.string().required('Username is required').max(40, 'Username is too long'),
         email: Yup.string().required('Email is required').email('Email is not valid'),
         password: Yup.string()
           .required('Password is required')
@@ -61,7 +66,6 @@ export default function Login({ handleSubmit }: Props): JSX.Element {
             }}
             name="email"
             autoComplete="email"
-            autoFocus
             helperText={touched.email ? errors.email : ''}
             error={touched.email && Boolean(errors.email)}
             value={values.email}
@@ -69,8 +73,30 @@ export default function Login({ handleSubmit }: Props): JSX.Element {
             className={classes.textField}
           />
           <TextField
+            id="username"
+            placeholder="Your name"
+            label={<Typography className={classes.label}>Name</Typography>}
+            fullWidth
+            margin="normal"
+            InputLabelProps={{
+              shrink: true,
+            }}
+            InputProps={{
+              classes: { input: classes.inputs },
+              disableUnderline: true,
+            }}
+            name="username"
+            autoComplete="username"
+            autoFocus
+            helperText={touched.username ? errors.username : ''}
+            error={touched.username && Boolean(errors.username)}
+            value={values.username}
+            onChange={handleChange}
+            className={classes.textField}
+          />
+          <TextField
             id="password"
-            placeholder="Your password"
+            placeholder="Create a password"
             label={<Typography className={classes.label}>Password</Typography>}
             fullWidth
             margin="normal"
@@ -80,7 +106,6 @@ export default function Login({ handleSubmit }: Props): JSX.Element {
             InputProps={{
               classes: { input: classes.inputs },
               disableUnderline: true,
-              endAdornment: <Typography className={classes.forgot}>Forgot?</Typography>,
             }}
             type="password"
             autoComplete="current-password"
@@ -92,11 +117,13 @@ export default function Login({ handleSubmit }: Props): JSX.Element {
           />
           <Box textAlign="center">
             <Button type="submit" size="large" variant="contained" color="primary" className={classes.submit}>
-              {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Login'}
+              {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Sign up'}
             </Button>
           </Box>
         </form>
       )}
     </Formik>
   );
-}
+};
+
+export default SignUpForm;
