--- client/src/pages/SignUp/useStyles.ts
@@ -1,50 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    backgroundColor: theme.palette.grey[200],
-    minHeight: '100vh',
-    '& .MuiInput-underline:before': {
-      borderBottom: '1.2px solid rgba(0, 0, 0, 0.2)',
-    },
-  },
-  authWrapper: {
-    display: 'flex',
-    alignItems: 'flex-start',
-    justifyContent: 'space-between',
-    flexDirection: 'column',
-    paddingTop: 23,
-  },
-  welcome: {
-    fontSize: 30,
-    paddingBottom: 20,
-    color: '#000000',
-    fontWeight: 700,
-    fontFamily: "'Open Sans'",
-  },
-  addMargin: {
-    marginTop: theme.spacing(6),
-    height: 'min-content',
-    [theme.breakpoints.up('xs')]: {
-      width: '100%',
-    },
-    [theme.breakpoints.up('sm')]: {
-      width: '85%',
-    },
-    [theme.breakpoints.up('md')]: {
-      width: '70%',
-    },
-    [theme.breakpoints.up('lg')]: {
-      width: '50%',
-    },
-  },
-  redirect: {
-    fontWeight: theme.typography.fontWeightBold,
-    margin: '2rem',
-    '& a': {
-      color: theme.palette.primary.main,
-    },
-  },
-}));
-
-export default useStyles;
